Stable file IDs across runs
These changes address https://github.com/tomakehurst/wiremock/issues/817.
A new option, file ID method, has been added.  It can be accessed on the
command line with --file-id-method.

The default file ID method is the existing method, RANDOM.

File ID methods:

* RANDOM
Randomly generated 5 digit alphanumeric ID.  This method generates a new
ID for every run, even if the HTTP request and response are the same.

* REQUEST_HASH
ID is hash based on the HTTP request.  Useful if you want to see if
responses for a given HTTP request are changing.    Caution: repeated
requests to the same endpoint will overwrite previous responses.

* RESPONSE_HASH
ID is hash based on the HTTP response.  Useful to minimize the number
of files created when two different HTTP requests produce identical
responses.

* REQUEST_RESPONSE_HASH
ID is a hash based on both the HTTP request and response.

Further work may be needed here to incorporate a scenario ID into the
request hash if the request is part of a scenario.